+++ LuaScriptReader/Properties/AssemblyInfo.cs
@@ -5,12 +5,12 @@ using System.Runtime.InteropServices;
 // General Information about an assembly is controlled through the following 
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("LuaScriptReader")]
+[assembly: AssemblyTitle("LuaScript")]
 [assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("LuaScriptReader")]
-[assembly: AssemblyCopyright("Copyright ©  2016")]
+[assembly: AssemblyProduct("LuaScript")]
+[assembly: AssemblyCopyright("Copyright © benmosheron@gmail.com 2016")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]